mock metodda gelen dizi kontrolu ve donen dizi yapildi.

+++ test/Base_Test/selectTests.h
@@ -179,7 +179,7 @@ TEST_F(TestSelect, _buildSelect)
     char            aid[]           = {0xA0, 0x00, 0x00, 0x00, 0x01};
     int             ret             = 0;
     unsigned char   actual[255]     = {0x00};
-    int             size            = 0;
+    unsigned long   size            = 0;
     unsigned char   expected[255]   = {0x00, 0xA4, 0x04, 0x00, 0x05, 0xA0, 0x00, 0x00, 0x00, 0x01};
     int             expectedSize    = 10;
 
@@ -212,7 +212,7 @@ TEST_F(TestSelect, _buildSelectPPSE)
     unsigned char   expected[255]   = {0x00, 0xA4, 0x04, 0x00, 0x0E,
                     '2','P','A','Y','.','S','Y','S','.','D','D','F','0','1'};
     int             expectedSize    = 19;
-    int             size            = 0;
+    unsigned long   size            = 0;
 
     ret = _buildSelectPpse(actual, &size);
